+++ actionview/test/template/form_helper_test.rb
@@ -2,6 +2,7 @@
 
 require "abstract_unit"
 require "controller/fake_models"
+require "active_support/core_ext/object/with"
 
 class FormHelperTest < ActionView::TestCase
   include RenderERBUtils
@@ -4097,6 +4098,14 @@ class FormHelperTest < ActionView::TestCase
     form_for(@post, data: { behavior: "stuff" }, remote: true) { }
     assert_match %r|data-behavior="stuff"|, @rendered
     assert_match %r|data-remote="true"|, @rendered
+  end
+
+  def test_form_for_with_configured_nest_html_attributes_within
+    ActionView::Helpers::TagHelper::TagBuilder.with nest_html_attributes_within: [:hx, "hx"] do
+      form_for(@post, html: { hx: { post: "/path" } }) { }
+
+      assert_dom "form[hx-post=?]", "/path"
+    end
   end
 
   def test_fields_for_returns_block_result